requirements.txt
ag2==0.7.6
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
autogen-core==0.4.8
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fast-depends==2.4.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jsonref==1.1.0
openai==1.65.3
opentelemetry-api==1.30.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pyautogen==0.7.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
